pyproject.toml
[project]
name = "python-aws-cmd"
version = "0.1.0"
description = ""
authors = [
    {name = "giorgi-kiknavelidze",email = "69247736+giorgi-kiknavelidze@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "boto3 (>=1.37.23,<2.0.0)",
    "types-boto3[full] (>=1.37.23,<2.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "typer (>=0.15.2,<0.16.0)"
]

[tool.poetry]
packages = [{include = "python_aws_cmd", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
